homework_8

+++ src/main/java/ru/otus/homework/service/BookServiceMongo.java
@@ -1,33 +1,23 @@
 package ru.otus.homework.service;
 
 import lombok.RequiredArgsConstructor;
-import org.hibernate.Hibernate;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.otus.homework.model.Author;
 import ru.otus.homework.model.Book;
-import ru.otus.homework.model.Comment;
-import ru.otus.homework.model.Genre;
-import ru.otus.homework.repository.AuthorRepositoryJpa;
-import ru.otus.homework.repository.BookRepositoryJpa;
-import ru.otus.homework.repository.CommentRepositoryJpa;
-import ru.otus.homework.repository.GenreRepositoryJpa;
+import ru.otus.homework.repository.BookRepositoryMongo;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 @ShellComponent
 @RequiredArgsConstructor
-public class BookServiceOnH2 implements BookService {
+public class BookServiceMongo implements BookService {
 
-    private final BookRepositoryJpa bookRepo;
-    private final AuthorRepositoryJpa authorRepo;
-    private final GenreRepositoryJpa genreRepo;
-    private final CommentRepositoryJpa commentRepo;
+    private final BookRepositoryMongo bookRepo;
+    private final CreateBookServiceConsole createBookService;
     private final IOService ioService;
 
     @Override
@@ -42,7 +32,7 @@ public class BookServiceOnH2 implements BookService {
     @Transactional
     @ShellMethod(value = "Add book", key = "add book")
     public String addBook() {
-        Book book = askAttributesCreateAndGetBook();
+        Book book = createBookService.askAttributesCreateAndGetBook();
         Book addedBook = bookRepo.save(book);
 
         return "Book " + addedBook.getTitle() + " added";
@@ -51,16 +41,14 @@ public class BookServiceOnH2 implements BookService {
     @Override
     @Transactional(readOnly = true)
     @ShellMethod(value = "Find book by id", key = "get book")
-    public String getBookById(long id) {
+    public String getBookById(String id) {
         Optional<Book> bookOptional = bookRepo.findById(id);
 
         if (bookOptional.isEmpty()) {
             ioService.out("Book not found");
             return null;
         }
-        List<Comment> comments = commentRepo.findAllByBookId(id);
-
-        return bookOptional.get() + ", comments: " + comments;
+        return bookOptional.get().toString();
     }
 
     @Override
@@ -81,17 +69,21 @@ public class BookServiceOnH2 implements BookService {
             ioService.out(String.format("Find several books: \n " +
                     "%s \n" +
                     "Please input id:", booksByTitle));
-            long id = Long.parseLong(ioService.read());
+            String id = ioService.read();
+
+            if (!bookRepo.existsById(id)) {
+                return "Book not found";
+            }
             bookRepo.deleteById(id);
         }
 
-        return "Book deleted :'(";
+        return "Book" + title + "deleted :'(";
     }
 
     @Override
     @Transactional
     @ShellMethod(value = "Delete book", key = "delete book id")
-    public String deleteBookById(long id) {
+    public String deleteBookById(String id) {
         if (!bookRepo.existsById(id)) {
             return "Book not found";
         }
@@ -119,50 +111,8 @@ public class BookServiceOnH2 implements BookService {
     @Override
     @Transactional(readOnly = true)
     @ShellMethod(value = "Get books by genre", key = "all genre")
-    public List<Book> getAllBooksByGenre(long id) {
-        Optional<Genre> optionalGenre = genreRepo.findById(id);
+    public List<Book> getAllBooksByGenre(String description) {
 
-        if (optionalGenre.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        List<Book> books = optionalGenre.get().getBooks();
-        Hibernate.initialize(books);
-
-        return books;
-    }
-
-    private Book askAttributesCreateAndGetBook() {
-        Book book = new Book();
-
-        ioService.out("Enter a book title");
-        String title = ioService.read();
-        ioService.out("Enter the first name of the author of book");
-        String firstName = ioService.read();
-        ioService.out("Enter the second name of the author of book");
-        String secondName = ioService.read();
-        ioService.out("Enter a book genre");
-        String genreDesc = ioService.read();
-        ioService.out("Enter a comment");
-        String commentAsString = ioService.read();
-        book.setTitle(title);
-
-        Author author = new Author();
-        author.setFirstName(firstName);
-        author.setSecondName(secondName);
-        authorRepo.save(author);
-        book.setAuthor(author);
-
-        Genre genre = new Genre();
-        genre.setDescription(genreDesc);
-        genreRepo.save(genre);
-        book.setGenre(genre);
-
-        Comment comment = new Comment();
-        comment.setText(commentAsString);
-        comment.setBook(book);
-        commentRepo.save(comment);
-
-        return book;
+        return bookRepo.findAllByGenre_Description(description);
     }
 }
